Split chapter and job modules into focused files
